fix(markdown): use shared tokenStyle utility with shiki's getTokenStyleObject
Combine both approaches: shared tokenStyle.ts uses getTokenStyleObject
from shiki (not @shikijs/core) to avoid pnpm resolution issues (#50),
while keeping a centralized utility to eliminate duplicated logic across
components.
