+++ packages/x-markdown/package.json
@@ -31,16 +31,12 @@
   "author": "",
   "license": "MIT",
   "peerDependencies": {
-    "@shikijs/core": "^3.0.0",
     "mermaid": "^11.0.0",
     "shiki": "^3.0.0",
     "shiki-stream": "^0.1.3",
     "vue": "^3.3.0"
   },
   "peerDependenciesMeta": {
-    "@shikijs/core": {
-      "optional": true
-    },
     "mermaid": {
       "optional": true
     },
@@ -74,7 +70,6 @@
     "unist-util-visit": "^5.0.0"
   },
   "devDependencies": {
-    "@shikijs/core": "^3.17.1",
     "@types/dompurify": "^3.0.0",
     "@types/hast": "^3.0.4",
     "@types/lodash-es": "^4.17.12",
