+++ packages/x-markdown/src/components/CodeBlock/SyntaxCodeBlock.vue
@@ -22,7 +22,7 @@
               v-else
               v-for="(token, j) in line"
               :key="j"
-              :style="getTokenStyle(token)"
+              :style="getTokenStyle(token, props.colorReplacements)"
               :class="{ 'x-md-animated-word': props.enableAnimate }"
               >{{ token.content }}</span
             >
@@ -35,9 +35,9 @@
 
 <script setup lang="ts">
 import { computed, type CSSProperties } from 'vue'
-import type { ThemedToken, BuiltinTheme } from 'shiki'
-import { getTokenStyleObject } from '@shikijs/core'
+import type { BuiltinTheme } from 'shiki'
 import { useHighlight } from '../../hooks/useHighlight'
+import { getTokenStyle } from '../../utils/tokenStyle'
 import type { SyntaxCodeBlockProps } from './types'
 
 defineOptions({
@@ -63,38 +63,6 @@ const { lines, preStyle } = useHighlight(code, {
   theme: actualTheme,
   colorReplacements: props.colorReplacements,
 })
-
-const applyColorReplacement = (color: string, replacements?: Record<string, string>) => {
-  if (!replacements) return color
-  return replacements[color.toLowerCase()] || color
-}
-
-const normalizeStyleKeys = (style: Record<string, string | number>): CSSProperties => {
-  const normalized: CSSProperties = {}
-  Object.entries(style).forEach(([key, value]) => {
-    const camelKey = key.replace(/-([a-z])/g, (_, char) => char.toUpperCase())
-    ;(normalized as Record<string, string | number>)[camelKey] = value
-  })
-  return normalized
-}
-
-const getTokenStyle = (token: ThemedToken): CSSProperties => {
-  const rawStyle = token.htmlStyle || getTokenStyleObject(token)
-  const baseStyle = normalizeStyleKeys(rawStyle)
-
-  if (!props.colorReplacements) return baseStyle
-
-  const style = { ...baseStyle }
-
-  if (style.color && typeof style.color === 'string') {
-    style.color = applyColorReplacement(style.color, props.colorReplacements)
-  }
-  if (style.backgroundColor && typeof style.backgroundColor === 'string') {
-    style.backgroundColor = applyColorReplacement(style.backgroundColor, props.colorReplacements)
-  }
-
-  return style
-}
 
 const showFallback = computed(() => !lines.value?.length)
